requirements.txt
flask
rembg
pillow
onnxruntime